Started homework2, make it work

+++ utils/io.py
@@ -1,3 +1,3 @@
 def read_matrix(filename, separator=" "):
     with open(filename, "r") as f:
-        return [list(map(int, line.split(separator))) for line in f]
+        return [list(map(float, line.split(separator))) for line in f]
